Beautification: Move layer helpers to make another module file
